コード整理

+++ ES Game Library/ES Game Library/MyClass/Data/MyAlgorithm.h
@@ -139,7 +139,7 @@ static Vector3 DirectionFromAngle(Vector3 angle)
 
 	Matrix front_matrix = matrix_x * matrix_y * matrix_z;
 
-	return Vector3_Normalize(Vector3(front_matrix._31, front_matrix._32, front_matrix._33));
+	return Vector3_Normalize(Vector3((int)front_matrix._31, (int)front_matrix._32, (int)front_matrix._33));
 }
 
 //! @fn  Path•ÏŠ·
